[Fix] TaskModal: 분리 전으로 롤백
