+++ src/components/modalInput/AssigneeSelect.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import RandomProfile from "@/components/common/RandomProfile";
+import RandomProfile from "../common/RandomProfile";
 
 interface User {
   id: number;
@@ -34,7 +34,7 @@ export default function AssigneeSelect({
   return (
     <div className="inline-flex flex-col items-start gap-2.5 w-full max-w-[520px]">
       {label && (
-        <p className="text-black3 font-medium text-[16px] sm:text-[18px]">
+        <p className="text-black3 font-medium text-[14px] sm:text-[18px]">
           {label}
           {required && <span className="text-[var(--primary)]"> *</span>}
         </p>
@@ -67,7 +67,7 @@ export default function AssigneeSelect({
                 </span>
               </>
             ) : (
-              <span className="font-normal text-[16px] sm:text-[18px] text-[var(--color-gray2)]">
+              <span className="font-normal text-[14px] sm:text-[16px] text-[var(--color-gray2)]">
                 담당자를 선택해 주세요
               </span>
             )}
