+++ src/components/modalInput/TaskModal.tsx
@@ -1,11 +1,14 @@
+import { useState, useMemo } from "react";
+import { useQuery, useQueryClient } from "@tanstack/react-query";
+import { createCard, editCard } from "@/api/card";
+import { getColumn } from "@/api/columns";
 import ModalInput from "@/components/modalInput/ModalInput";
 import ModalTextarea from "@/components/modalInput/ModalTextarea";
 import ModalImage from "@/components/modalInput/ModalImage";
 import TextButton from "@/components/modalInput/TextButton";
 import StatusSelect from "@/components/modalInput/StatusSelect";
 import AssigneeSelect from "@/components/modalInput/AssigneeSelect";
-import { useTaskForm } from "@/hooks/useTaskForm";
-import { useColumnStatus } from "@/hooks/useColumnStatus";
+import { toast } from "react-toastify";
 
 interface TaskModalProps {
   mode?: "create" | "edit";
@@ -21,7 +24,7 @@ interface TaskModalProps {
   }[];
   columnId: number;
   dashboardId: number;
-  cardId?: number;
+  cardId?: number; // 수정 모드일 때만 사용
 }
 
 export interface TaskData {
@@ -34,6 +37,12 @@ export interface TaskData {
   image?: string;
 }
 
+interface ColumnType {
+  id: number;
+  title: string;
+  status: string;
+}
+
 export default function TaskModal({
   mode = "create",
   onClose,
@@ -44,33 +53,107 @@ export default function TaskModal({
   dashboardId,
   cardId,
 }: TaskModalProps) {
-  const updatedColumnId = useColumnStatus(
-    dashboardId,
-    columnId,
-    initialData.status || ""
-  );
+  const queryClient = useQueryClient();
 
-  const { formData, handleChange, isFormValid, handleSubmit } = useTaskForm({
-    mode,
-    initialData,
-    members,
-    dashboardId,
-    columnId,
-    cardId,
-    updatedColumnId,
-    onClose,
-    onSubmit,
+  const [formData, setFormData] = useState<TaskData>({
+    status: initialData.status || "",
+    assignee: initialData.assignee || "",
+    title: initialData.title || "",
+    description: initialData.description || "",
+    deadline: initialData.deadline ?? "",
+    tags: initialData.tags || [],
+    image: initialData.image || "",
   });
+
+  const { data: columns = [] } = useQuery<ColumnType[]>({
+    queryKey: ["columns", dashboardId],
+    queryFn: () => getColumn({ dashboardId, columnId }),
+  });
+
+  const matchedColumn = useMemo(() => {
+    if (!columns.length) return undefined;
+    return columns.find((col) => col.title === formData.status);
+  }, [columns, formData.status]);
+
+  const updatedColumnId = matchedColumn?.id ?? columnId;
+
+  const handleChange = (field: keyof TaskData, value: string | string[]) => {
+    setFormData((prev) => ({
+      ...prev,
+      [field]: value,
+    }));
+  };
+
+  const isFormValid =
+    formData.assignee &&
+    formData.status &&
+    formData.title.trim() &&
+    formData.description.trim();
+
+  const handleSubmit = async () => {
+    try {
+      const selectedAssignee = members.find(
+        (m) => m.nickname === formData.assignee
+      );
+      const assigneeUserId = selectedAssignee?.userId;
+
+      if (!assigneeUserId) {
+        toast.error("담당자를 선택해 주세요.");
+        return;
+      }
+
+      if (mode === "create") {
+        await createCard({
+          assigneeUserId,
+          dashboardId,
+          columnId: updatedColumnId,
+          title: formData.title,
+          description: formData.description,
+          dueDate: formData.deadline.trim() ? formData.deadline : undefined,
+          tags: formData.tags,
+          imageUrl: formData.image || undefined,
+        });
+        toast.success("카드가 생성되었습니다.");
+      } else {
+        if (!cardId) {
+          toast.error("카드 ID가 없습니다.");
+          return;
+        }
+
+        await editCard(cardId, {
+          assigneeUserId,
+          columnId: updatedColumnId,
+          title: formData.title,
+          description: formData.description,
+          dueDate: formData.deadline.trim() ? formData.deadline : undefined,
+          tags: formData.tags,
+          imageUrl: formData.image || undefined,
+        });
+        queryClient.invalidateQueries({ queryKey: ["cards"] });
+        toast.success("카드가 수정되었습니다.");
+      }
+
+      onSubmit?.(formData);
+      onClose();
+    } catch (err) {
+      console.error("카드 처리 실패:", err);
+      toast.error(`카드 ${mode === "edit" ? "수정" : "생성"}에 실패했습니다.`);
+    }
+  };
+
   return (
     <div className="fixed inset-0 flex items-center justify-center bg-black/35 z-50">
-      <div className="sm:w-[584px] w-[320px] h-[calc(var(--vh)_*_90)] rounded-lg bg-white p-4 sm:p-8 shadow-lg flex flex-col gap-4 sm:gap-8 overflow-y-auto">
-        <h2 className="font-bold text-black3 text-[18px] sm:text-[24px]">
+      <div className="sm:w-[584px] w-[327px] h-[calc(var(--vh)_*_90)] rounded-lg bg-white p-4 sm:p-8 shadow-lg flex flex-col gap-4 sm:gap-8 overflow-y-auto">
+        <h2 className="font-bold text-black3 text-[16px] sm:text-[24px]">
           {mode === "edit" ? "할 일 수정" : "할 일 생성"}
         </h2>
 
         <div className="flex flex-col gap-4 sm:gap-8">
           {/* 상태 및 담당자 */}
-          <div className="flex flex-col sm:flex-row gap-4 text-black3 font-normal text-[16px] sm:text-[18px]">
+          <div
+            className="flex flex-col sm:flex-row gap-4
+          text-black3 font-normal text-[14px] sm:text-[16px]"
+          >
             <StatusSelect
               label="상태"
               value={formData.status}
@@ -128,7 +211,8 @@ export default function TaskModal({
             color="third"
             buttonSize="md"
             onClick={onClose}
-            className="sm:w-[256px] w-[144px] h-[54px] border border-[var(--color-gray3)] bg-white text-[var(--color-gray1)] font-16m rounded-lg cursor-pointer"
+            className="sm:w-[256px] w-[144px] h-[54px] border border-[var(--color-gray3)] bg-white
+            text-[var(--color-gray1)] font-16m rounded-lg cursor-pointer"
           >
             취소
           </TextButton>
